+++ phoneProfilesPlus/src/main/java/sk/henrichg/phoneprofilesplus/ApplicationsDialogPreferenceX.java
@@ -599,15 +599,18 @@ public class ApplicationsDialogPreferenceX extends DialogPreference {
         popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
 
             public boolean onMenuItemClick(android.view.MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.applications_pref_dlg_item_menu_edit:
-                        startEditor(application);
-                        return true;
-                    case R.id.applications_pref_dlg_item_menu_delete:
-                        deleteApplication(application);
-                        return true;
-                    default:
-                        return false;
+                int itemId = item.getItemId();
+                if (itemId == R.id.applications_pref_dlg_item_menu_edit) {
+                    startEditor(application);
+                    return true;
+                }
+                else
+                if (itemId == R.id.applications_pref_dlg_item_menu_delete) {
+                    deleteApplication(application);
+                    return true;
+                }
+                else {
+                    return false;
                 }
             }
         });
